handle A/B from different LUNs
Some Qualcomm devices rely on each slot partition set to exist on
separate LUNs. Add the logic in boot_ctl_set_active_slot_for_partitions
to handle A and B being on different LUNs.

While doing so, move the struct gpt_disk down in the function to manage
its live cycle there since it is not used after the function is invoked.

Signed-off-by: Eric Chanudet <echanude@redhat.com>