HACK: sa8540p-ride: Only set boot_a/b partitions
On these Qualcomm SoC the boot status is only written to the boot
partitions.

Since it is not the job of this tool to figure out what SoC it's running
on, nor how its storage was provisioned, it would be preferable to have
the partition list passed as argument by whatever invokes it.
In turn, that would mandate some more elaborate option parsing.

Signed-off-by: Eric Chanudet <echanude@redhat.com>